ventas_app/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.3
SQLAlchemy==1.3.24
python-dotenv==0.19.1